More tweaks
Use proper type for notification handler, it's not UIApplicationShortcutItem. Name it better. There's no need to use mutable dictionary (even if, prefer mutableCopy method). If we decide to log output, let's use React builtins and prefix with `[libraryName]` for clarity.

Currently cold launching throws for some reason, investigating. The rest of the code works as expected.